Apply card-based theme and model/controller updates
- Updated Home page and _Layout to match card-based landing page theme
- Updated Customer views (Index, Create, Edit) to card-based theme
- Updated Product views (Index, Create, Edit) to card-based theme
- Updated CustomerController to fix edit issue preventing updates to customer info
- Changed decimal fields to double in Product and Order models and ProductController
- Applied consistent green accent color across homepage, customer, and product pages

+++ ABCRetailers/Controllers/ProductController.cs
@@ -35,7 +35,7 @@ namespace ABCRetailers.Controllers
             {
                 _logger.LogInformation("Raw price from form: '{PriceForValue}'", priceFormValue.ToString());
 
-                if (decimal.TryParse(priceFormValue, out var parsedPrice))
+                if (double.TryParse(priceFormValue, out var parsedPrice))
                 {
                     product.Price = parsedPrice;
                     _logger.LogInformation("Successfully parsed price: {Price}", parsedPrice);
@@ -103,7 +103,7 @@ namespace ABCRetailers.Controllers
             // Manual price parsing for edit too
             if (Request.Form.TryGetValue("Price", out var priceFormValue))
             {
-                if (decimal.TryParse(priceFormValue, out var parsedPrice))
+                if (double.TryParse(priceFormValue, out var parsedPrice))
                 {
                     product.Price = parsedPrice;
                     _logger.LogInformation("Edit: Successfully parsed price: {Price}", parsedPrice);
